07/10/2023 - Added ft_lstclear, ft_lstiter, ft_lstmap and Makefile. Every file is compliant with norminette.

+++ Common_Core/Libft/ft_functions/ft_strlcat.c
@@ -11,7 +11,8 @@
 /* ************************************************************************** */
 
 #include <stdio.h>
-static unsigned int	cat(char *dest, char const *src, size_t size, unsigned int j);
+
+static size_t	cat(char *dest, char const *src, size_t size, unsigned int j);
 
 unsigned int	ft_strlcat(char *dest, char const *src, size_t size)
 {
@@ -35,7 +36,7 @@ unsigned int	ft_strlcat(char *dest, char const *src, size_t size)
 	return (j);
 }
 
-static unsigned int	cat(char *dest, char const *src, size_t size, unsigned int j)
+static size_t	cat(char *dest, char const *src, size_t size, unsigned int j)
 {
 	while (*src != '\0')
 	{
